Restructure pages
Magazine 2 now contains SoundLink information, originally meant to be in mag 3. Pages 101 and 201 are now in separate sub-pages, rather than in one big carousel.
NOTE TO SCHEDULERS: If the Broadcast Station has more that 3 games on any specific day, don't forget to set the Transmit flag on p101-02. Else it won't show up.